Version bump and associated cleanups
This commit increases the version to 0.0.3 and performs multiple
cleanups to prepare the new relase

We rename the module GitManager.main to GitManager.__main__ and add a
bit of utility code to allow calls of the form `python -m GitManager ...`.

Inside of setup.py we use the find_packages() command instead of listing
all packages manually.

We remove support for VCs that are not git and add a new
abstraction to allow easier usage thereof.

We updated the format module, in particular we reworked the
path() function by splitting it up into a relative and absolute version.

We also added unit tests for the format and the new run module.

Finally we also made minor other improvements such as updating the year
the copyright year and making other files PEP8 compliant.