Improve TerminalLine() behaviour on non-ttys
Previously, if the TerminalLine() helper class was called on a non-ttys
file, all lines were printed. This resulted in lines that were not
visible in the final output of a tty terminal.

This commit improves upon the behaviour by having calls to append() and
write() cache the current content of the TerminalLine(). Then upon
inserting a linebreak or calling the flush function, cached older lines
are written out.

This results in the output on non-tty files looking the same as final
output on tty terminals.